Pull statistics out into separate component
