+++ src/App.tsx
@@ -2,11 +2,12 @@ import * as React from 'react';
 import './App.css';
 import {
   addCornDense, addCornSparse, addTrapCropDense, addTrapCropSparse,
-  addWormsSparse, getCornStats
+  addWormsSparse
 } from './corn-model';
 import { worm } from './species/rootworm';
-import { Environment, Events, Species } from './populations';
+import { Species } from './populations';
 import { Attribution } from './components/attribution';
+import { SimulationStatistics } from './components/simulation-statistics';
 import PopulationsModel from './components/populations-model';
 import { forEach } from 'lodash';
 
@@ -63,16 +64,6 @@ interface IAppProps {
 }
 
 interface IAppState {
-  initialCorn: number;
-  initialTrap: number;
-  initialWorms: number;
-  dayFirstCornEaten: number | null;
-  totalCorn: number;
-  totalTrap: number;
-  totalWorm: number;
-  infectedCorn: number;
-  harvestCorn: number;
-  simulationDay: number;
   // store as strings during editing
   wormEatingDistance: string;
   wormEnergy: string;
@@ -87,16 +78,7 @@ interface IAppState {
 class App extends React.Component<IAppProps, IAppState> {
 
   public state: IAppState = {
-    initialCorn: 0,
-    initialTrap: 0,
-    initialWorms: 0,
-    dayFirstCornEaten: null,
-    totalCorn: 0,
-    totalTrap: 0,
-    totalWorm: 0,
-    infectedCorn: 0,
-    harvestCorn: 0,
-    simulationDay: 0,
+
     wormEatingDistance: "",
     wormEnergy: "",
     wormMetabolism: "",
@@ -117,35 +99,6 @@ class App extends React.Component<IAppProps, IAppState> {
         traitState[spec.stateName] = String(defaultValue);
       }
       this.setState(traitState as any);
-    });
-
-    Events.addEventListener(Environment.EVENTS.STEP, (evt: any) => {
-      let { dayFirstCornEaten } = this.state;
-      const { countCorn, countTrap, countWorm, infected, simulationDay } = getCornStats();
-      const actualDay = Math.trunc(simulationDay / 3);
-      if ((this.state.dayFirstCornEaten == null) && (countCorn < this.state.initialCorn)) {
-        dayFirstCornEaten = actualDay;
-      }
-      this.setState({
-        totalCorn: countCorn, totalTrap: countTrap, totalWorm: countWorm,
-        dayFirstCornEaten, infectedCorn: infected, simulationDay: actualDay
-      });
-    });
-
-    Events.addEventListener(Environment.EVENTS.START, (evt: any) => {
-      const { countCorn, countTrap, countWorm, infected } = getCornStats();
-      this.setState({
-        initialCorn: countCorn, initialTrap: countTrap, initialWorms: countWorm,
-        dayFirstCornEaten: null, infectedCorn: infected
-      });
-    });
-
-    Events.addEventListener(Environment.EVENTS.RESET, (evt: any) => {
-      this.setState({
-        initialCorn: 0, initialTrap: 0, initialWorms: 0,
-        totalCorn: 0, totalTrap: 0, totalWorm: 0,
-        dayFirstCornEaten: null, infectedCorn: 0, simulationDay: 0
-      });
     });
   }
 
@@ -284,16 +237,7 @@ class App extends React.Component<IAppProps, IAppState> {
               </div>
             </div>
           </div>
-          <div className="section stats">
-            <h4>Statistics</h4>
-            <div><span>Day: </span><span>{this.state.simulationDay}</span></div>
-            <div><span>Corn planted: </span><span>{this.state.initialCorn}</span></div>
-            <div><span>Corn remaining: </span><span>{this.state.totalCorn}</span></div>
-            <div><span>First corn eaten: </span><span>{this.state.dayFirstCornEaten}</span></div>
-            <div><span>Trap planted: </span><span>{this.state.initialTrap}</span></div>
-            <div><span>Trap remaining: </span><span>{this.state.totalTrap}</span></div>
-            <div><span>Worms: </span><span>{this.state.totalWorm}</span></div>
-          </div>
+          <SimulationStatistics />
         </div>
         <Attribution />
       </div>
